requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.3
pydantic-settings==2.3.1

# Base de datos (SQLite async)
sqlalchemy==2.0.31
aiosqlite==0.20.0

# Auth (validación del JWT de Supabase)
python-jose[cryptography]==3.3.0
httpx==0.27.0

# PDF
reportlab==4.2.2

# Utilidades
python-dotenv==1.0.1

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
